refactor useUI
